refactor(cloudwatch-anomaly-detector): rearrange struct location, rename pointers to standard
